requirements.txt
pre-commit>=2.16.0
pre-commit-hooks>=4.1.0
pylint>=2.12.2
black>=21.12b0
docformatter>=1.4.0
mistune>=2.0.5

fastapi>=0.110.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
uvicorn>=0.20.0

